python-services/requirements_vector.txt
# Requirements for StackLensAI Vector Database Service
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pydantic==2.4.2
python-multipart==0.0.6
scikit-learn==1.3.0
torch==2.1.0
transformers==4.34.0

# Optional: For GPU support (uncomment if you have CUDA)
# faiss-gpu==1.7.4
# torch==2.1.0+cu118 --index-url https://download.pytorch.org/whl/cu118
